% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cry2-mspct.r
\docType{data}
\name{CRY2.mspct}
\alias{CRY2.mspct}
\title{CRY2 Absorbance spectra.}
\format{A \code{filter.mspct} with two member \code{filter.spct} objects each
  with 100 or 200 rows and 2 numeric variables, \code{w.length} and \code{A}}
\description{
A dataset containing the wavelengths at an arbitrary nm interval. Tabulated
values for the in vitro absorbance spectrum of CRY2. The data were digitized
from Figure 1.B, curve "dark adapted sample", and curve "irradiated with blue
light (450 nm, 50 umol m-2 s-1) during 30 min" in Banerjee et al. (2007).
}
\details{
The variables of the member spectra are as follows:

\itemize{ \item w.length (nm) \item A (spectral absorbance) }
}
\references{
Banerjee, R., Schleicher, E., Meier, S., Viana, R. M., Pokorny,
  R., Ahmad, M., … Batschauer, A. (2007). The signaling state of Arabidopsis
  cryptochrome 2 contains flavin semiquinone. J Biol Chem, 282(20),
  14916–14922. doi:10.1074/jbc.M700616200
}
\keyword{datasets}

