% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D-water-CO2.R
\name{D_water}
\alias{D_water}
\alias{D_CO2}
\title{Diffusion coefficients in air}
\usage{
D_water(temperature)

D_CO2(temperature)
}
\arguments{
\item{temperature}{numeric Temperature [\eqn{^{\circ}C}], in the range -7
to 47 \eqn{^{\circ}C}.}
}
\value{
A numeric vector of the same length as \code{temperature} containing
   diffusion coefficients expressed in \eqn{m^2 s^{-1}}. \code{NA} is
  returned silently for off-range arguments passed to \code{temperature}.
}
\description{
Diffusion coefficients of water and carbon dioxide in excess air as a
function of temperature.
}
\details{
Implemented as interpolating splines from tabulated data.
}
\examples{
D_water(23)
D_water(c(10, 12, 15, 100))

D_CO2(11)

}
\references{
Monteith, J. L. and Unsworth M. H. (2008) Principles of Environmental
  Physics (3 ed.) Academic Press-Elsevier. ISBN: 978-0-12-505103-3. See
  Table A.3.
}
