% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs-anatomy.R
\name{gs_c_from_gs_w}
\alias{gs_c_from_gs_w}
\alias{gs_w_from_gs_c}
\alias{gs_vol2mol}
\alias{gs_mol2vol}
\title{Convert stomatal conductance}
\usage{
gs_c_from_gs_w(gs_w, temperature = 21)

gs_w_from_gs_c(gs_c, temperature = 21)

gs_vol2mol(gs, temperature = 21, pressure = 101300)

gs_mol2vol(gs_c, temperature = 21, pressure = 101300)
}
\arguments{
\item{gs_w, gs_c, gs}{numeric Stomatal conductance to water vapour and carbon
dioxide [\eqn{m^2 s^{-1}}].}

\item{temperature}{numeric Temperature [\eqn{^{\circ}C}], in the range -7
to 47 \eqn{^{\circ}C}.}

\item{pressure}{numeric Pressure of the gas, e.g., atmospheric pressure [Pa].}
}
\value{
A numeric vector of stomatal conductance values expressed as
  \eqn{m^2 s^{-1}}.
}
\description{
Convert between stomatal conductances to water vapour and carbon dioxide.
}
\details{
These conversions are based on the diffusion coefficients, both of
  which are looked up based on the temperature.
}
\examples{
# example code
gs_c_from_gs_w(1/120)
gs_vol2mol(1/120)

}
