% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs-anatomy.R
\name{molar_vol}
\alias{molar_vol}
\title{Molar volume of an ideal gas}
\usage{
molar_vol(temperature = 20, pressure = 101300)
}
\arguments{
\item{temperature}{numeric Temperature of the gas [\eqn{^{\circ}C}].}

\item{pressure}{numeric Pressure of the gas, e.g., atmospheric pressure [Pa].}
}
\value{
A numeric vector with the molar volume of a gas expressed in
  \eqn{m^3} per mole.
}
\description{
Compute the molar volume of an ideal gas from temåerature and pressure.
}
\details{
The ideal gas equation, \eqn{V = R * T / P}, is used to compute the
returned value. In the equation \eqn{R} is the gas constant
(\eqn{8.314 m^3 Pa mol^{-1} K^{-1}}), \eqn{T} temperature expressed
in degrees kelvin
and \eqn{P} the pressure expressed in pascals. The argument passed to
\code{temperature}, expressed in \eqn{^{\circ}C}, is re-expressed in
\eqn{K} for the computation.
}
\examples{
molar_vol()
molar_vol(0)
molar_vol(0, 90e3)

}
