% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pfr-P-ratio-mono.R
\name{Pfr_P_ratio_mono}
\alias{Pfr_P_ratio_mono}
\title{Calculation of Pfr:Ptot ratio for monochromatic radiation.}
\usage{
Pfr_P_ratio_mono(w.length)
}
\arguments{
\item{w.length}{a single value or a vector of wavelengths (nm)}
}
\value{
a single value or a vector of numeric values giving the unitless
  ratio
}
\description{
Ratio between Pr and Ptot (photoequilibrium)
for Type I Phytochrome.
}
\section{Warning!}{

Exact only for narrow band light sources!
}

\references{
Mancinelli, A.L. (1994) The physiology of phytochrome action. In
Photomorphogenesis in plants, 2nd edition. R.E. Kendrick and G.H.M.
Kronenberg, eds. Kluwer Academic Publishers, Dordrecht, pp. 211-269. ISBN
978-0-7923-2551-2 (print), 978-94-011-1884-2 (on-line).
\doi{10.1007/978-94-011-1884-2_10}
}
\keyword{internal}
