% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pfr-Ptot.R
\name{Pfr_Ptot}
\alias{Pfr_Ptot}
\alias{Pfr_Ptot.default}
\alias{Pfr_Ptot.numeric}
\alias{Pfr_Ptot.source_spct}
\alias{Pfr_Ptot.source_mspct}
\title{Calculate phytochrome photoequilibrium}
\usage{
Pfr_Ptot(x, ...)

\method{Pfr_Ptot}{default}(x, ...)

\method{Pfr_Ptot}{numeric}(x, spct.out = length(x) > 20, ...)

\method{Pfr_Ptot}{source_spct}(x, return.tb = !is.null(attr2tb), attr2tb = NULL, ..., na.rm = FALSE)

\method{Pfr_Ptot}{source_mspct}(
  x,
  return.tb = TRUE,
  attr2tb = NULL,
  ...,
  na.rm = FALSE,
  idx = "spct.idx",
  .parallel = FALSE,
  .paropts = NULL
)
}
\arguments{
\item{x}{an R object. A \code{numeric} vector}

\item{...}{currently ignored.}

\item{spct.out}{logical Flag indicating if the returned object should be of
class \code{response_spct} instead of \code{numeric}.}

\item{return.tb}{logical If \code{TRUE} force return of a data frame for a
single spectrum, to match the returned class for collections of spectra.}

\item{attr2tb}{character vector, see
\code{\link[photobiology]{add_attr2tb}()} the syntax for \code{attr2tb}
passed as is to formal parameter \code{col.names}.}

\item{na.rm}{logical. If \code{TRUE} \code{link[stats]{na.omit}} is first
called on \code{x}.}

\item{idx}{character Name of the column with the names of the members of the
collection of spectra.}

\item{.parallel}{if TRUE, apply function in parallel, using parallel backend
provided by foreach}

\item{.paropts}{a list of additional options passed into the foreach function
when parallel computation is enabled. This is important if (for example)
your code relies on external data or packages: use the .export and
.packages arguments to supply them so that all cluster nodes have the
correct environment set up for computing.}
}
\value{
When input is spectral data, a numeric vector of values
  giving the \eqn{P_{fr} / P_{tot}} or a data frame are returned, possibly
  with attributes extracted from the spectral objects as additional
  columns. When input is a numeric vector of wavelengths, the returned
  object is either a numeric vector or a \code{generic_spct} object with
  giving the \eqn{P_{fr} / P_{tot}} for each wavelength in the input.
}
\description{
Calculate the phytochrome photoequilibrium for monochromatic light from its
wavelength or from a spectrum expressed as spectral irradiance.
}
\details{
The calculations are based on data describing the photochemical
  constants for the plant photoreceptor phytochrome measured \emph{in vitro}
  and available for wavelengths in the range 380 nm to 770 nm as published by
  Mancinelli (1994). For reliable estimates of \eqn{P_{fr} / P_{tot}} from
  spectral irradiance, the spectrum should cover all these wavelengths with
  reasonably high wavelength resolution.

  Two approaches are possible, using wavelength for monochromatic radiation
  or spectral irradiance data. Spectral irradiance data are accepted either
  as individual spectra or collections of spectra.
}
\note{
If you use these data in a publication, please cite also the original
  source as given under references.
}
\examples{
# monochromatic light
Pfr_Ptot(620) # one wavelength in nm
Pfr_Ptot(c(570, 600, 630, 660, 690, 735, 760)) # six wavelengths
# spectral irradiance
Pfr_Ptot(sun.spct) # one spectrum
Pfr_Ptot(sun.spct, attr2tb = "when.measured")
Pfr_Ptot(sun_evening.spct) # five spectra
Pfr_Ptot(sun_evening.mspct)
Pfr_Ptot(sun_evening.mspct, attr2tb = "when.measured")

}
\references{
Mancinelli, A.L. (1994) The physiology of phytochrome action. In
Photomorphogenesis in plants, 2nd edition. R.E. Kendrick and G.H.M.
Kronenberg, eds. Kluwer Academic Publishers, Dordrecht, pp. 211-269. ISBN
978-0-7923-2551-2 (print), 978-94-011-1884-2 (on-line).
\doi{10.1007/978-94-011-1884-2_10}
}
\seealso{
Other phytochrome-related functions and data:
\code{\link{PHYs.mspct}},
\code{\link[=Pfr_Ptot_R_FR]{Pfr_Ptot_R_FR()}},
\code{\link[=Phy_Sigma]{Phy_Sigma()}},
\code{\link[=Phy_Sigma_FR]{Phy_Sigma_FR()}},
\code{\link[=Phy_Sigma_R]{Phy_Sigma_R()}},
\code{\link[=Phy_reaction_rates]{Phy_reaction_rates()}}
}
\concept{phytochrome-related functions and data}
