% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant-qq-ratios.r
\name{photon ratios}
\alias{photon ratios}
\alias{R_FR}
\alias{B_G}
\alias{UVB_UV}
\alias{UVB_UVA}
\alias{UVA_UV}
\alias{UVAlw_UV}
\alias{UVAsw_UV}
\alias{UV_PAR}
\alias{UVB_PAR}
\alias{UVA_PAR}
\alias{UVA1_UV}
\alias{UVA2_UV}
\alias{UVA2_UVA}
\title{Calculate photon ratios from spectral irradiance}
\usage{
R_FR(spct, std = "Smith20", use.cached.mult = FALSE, use.hinges = TRUE, ...)

B_G(spct, std = "Sellaro", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UVB_UV(spct, std = "ISO", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UVB_UVA(spct, std = "ISO", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UVA_UV(spct, std = "ISO", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UVAlw_UV(spct, std = "plants", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UVAsw_UV(spct, std = "plants", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UV_PAR(spct, std = "ISO", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UVB_PAR(spct, std = "ISO", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UVA_PAR(spct, std = "ISO", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UVA1_UV(spct, std = "CIE", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UVA2_UV(spct, std = "CIE", use.cached.mult = FALSE, use.hinges = TRUE, ...)

UVA2_UVA(spct, std = "CIE", use.cached.mult = FALSE, use.hinges = TRUE, ...)
}
\arguments{
\item{spct}{an object of class "source.spct" or an object of class
"source.mspct" containing one of more spectra.}

\item{std}{character Name of the variants of the two waveband definitions to
use.}

\item{use.cached.mult}{logical indicating whether multiplier values should be
cached between calls.}

\item{use.hinges}{logical indicating whether to use hinges to reduce
interpolation errors.}

\item{...}{named arguments to be forwarded to \code{q_ratio} methods.}
}
\value{
When \code{spct} contains a single spectrum, a single named numeric
  dimensionless value giving a photon ratio, with name constructed from the
  name of the wavebands, with "(q:q)" appended is returned. When \code{spct}
  contains multiple spectra, either in long form or as a collection of
  spectral objects, the returned object is a data frame with a factor
  identifying the spectra and a numeric variable with the numeric values of
  the ratio.
}
\description{
These functions return photon ratios between two different wavebands computed
from one or more light source spectra.
}
\details{
These functions are convenience wrappers on calls to method
  \code{\link[photobiology]{q_ratio}} with specific waveband definitions from
  package \code{\link[photobiologyWavebands]{photobiologyWavebands}}. To
  compute other photon ratios call method \code{\link[photobiology]{q_ratio}}
  with predefined or ad hoc \code{\link[photobiology]{waveband}} definitions.

  The returned value is the ratio between two photon irradiances (or two
  photon fluence values) each integrated over the range of wavelengths in a
  waveband definition, which can differ in wavelength extent. Some ratios
  are defined for non-overlapping ranges of wavelengths (e.g., R:FR photon
  ratio) while others are defined for overlapping ranges of wavelengths
  (e.g., UVB:UV, which will never exceed 1 in value).
}
\examples{
# default, one spectrum
R_FR(sun.spct)

# default, multiple spectra
R_FR(sun_evening.spct)
R_FR(sun_evening.mspct)

# different waveband definitions
R_FR(sun.spct, std = "Smith10")
R_FR(sun.spct, std = "Smith20")
R_FR(sun.spct, std = "Sellaro")
R_FR(sun.spct, std = "Apogee")

B_G(sun.spct)

UVB_UV(sun.spct)

UVB_UVA(sun.spct)

UVA_UV(sun.spct)

UVAlw_UV(sun.spct)

UVAsw_UV(sun.spct)

UV_PAR(sun.spct)

UVB_PAR(sun.spct)

UVA_PAR(sun.spct)

UVA1_UV(sun.spct)

UVA2_UV(sun.spct)

UVA2_UVA(sun.spct)

}
\seealso{
Ratios are computed with \code{\link[photobiology]{q_ratio}} with
  \code{\link[photobiology]{waveband}} objects as input. In the table below
  the wavebands and default for \code{std} used to compute the photon ratios
  in each of the functions are listed and linked to the respective help
  pages.

  \tabular{llll}{
     \strong{Function}  \tab \strong{Numerator} \tab \strong{Denominator} \tab \strong{Default} \code{std} \cr
     \code{R_FR}  \tab \code{\link[photobiologyWavebands]{Red}} \tab \code{\link[photobiologyWavebands]{Far_red}} \tab "Smith20" \cr
     \code{B_G}   \tab \code{\link[photobiologyWavebands]{Blue}} \tab \code{\link[photobiologyWavebands]{Green}} \tab "Sellaro" \cr
     \code{UVB_UV}  \tab \code{\link[photobiologyWavebands]{UVB}} \tab \code{\link[photobiologyWavebands]{UV}} \tab "ISO" \cr
     \code{UVB_UVA}  \tab \code{\link[photobiologyWavebands]{UVB}} \tab \code{\link[photobiologyWavebands]{UVA}} \tab "ISO" \cr
     \code{UVA_UV}  \tab \code{\link[photobiologyWavebands]{UVA}} \tab \code{\link[photobiologyWavebands]{UV}} \tab "ISO" \cr
     \code{UVAlw_UV}  \tab \code{\link[photobiologyWavebands]{UVAlw}} \tab \code{\link[photobiologyWavebands]{UV}} \tab "plants" \cr
     \code{UVAsw_UV}  \tab \code{\link[photobiologyWavebands]{UVAsw}} \tab \code{\link[photobiologyWavebands]{UV}} \tab "plants" \cr
     \code{UV_PAR}  \tab \code{\link[photobiologyWavebands]{UV}} \tab \code{\link[photobiologyWavebands]{PAR}} \tab "ISO" \cr
     \code{UVB_PAR}  \tab \code{\link[photobiologyWavebands]{UVB}} \tab \code{\link[photobiologyWavebands]{PAR}} \tab "ISO" \cr
     \code{UVA_PAR}  \tab \code{\link[photobiologyWavebands]{UVA}} \tab \code{\link[photobiologyWavebands]{PAR}} \tab "ISO" \cr
     \code{UVA1_UV}  \tab \code{\link[photobiologyWavebands]{UVA1}} \tab \code{\link[photobiologyWavebands]{UV}} \tab "CIE" \cr
     \code{UVA2_UV}  \tab \code{\link[photobiologyWavebands]{UVA2}} \tab \code{\link[photobiologyWavebands]{UV}} \tab "CIE" \cr
     \code{UVA2_UVA}  \tab \code{\link[photobiologyWavebands]{UVA2}} \tab \code{\link[photobiologyWavebands]{UVA}} \tab "CIE" \cr
  }
}
