% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phots-mspct.r
\docType{data}
\name{PHOTs.mspct}
\alias{PHOTs.mspct}
\title{PHOT1 and PHOT2 absorbance spectra.}
\format{A \code{filter_mspct} with five member \code{filter_spct} objects each
  with 300 rows and 2 numeric variables, \code{w.length} and \code{A}}
\description{
A dataset containing the wavelengths at an arbitrary nm interval. Tabulated
values for the in vitro absorbance spectrum of PHOT1 LOV2 domain for
fluorescence yield of PHOT1 and PHOT2 from Arabidopsis measured in vitro.
PHOT1 fluorescenve yield data were digitized from figure 1a curve
"LOV1 + LOV2 (WT) and PHOT2 fluorescenve yield data were digitized from figure
7a curve "LOV1 + LOV2 (WT) in Christie et al. (2002). PHOT1 LOV2, dark
adapted, spectral absorbance data were digitized from figure 3, black
curve and PHOT1 LOV2, blue-light adapted spectral absorbance data were
digitized from figure 3, blue curve in Christie et al. (2015).
}
\details{
The variables of the member spectra are as follows:

\itemize{ \item w.length (nm) \item A (spectral absorbance) }
}
\note{
If you use these data in a publication, please cite also the original
  source as given under references.
}
\references{
CHRISTIE, John M., SWARTZ, Trevor E., BOGOMOLNI, Roberto A., BRIGGS, Winslow
   R. (2002) Phototropin LOV domains exhibit distinct roles in regulating
   photoreceptor function. The Plant Journal 32(2):205-219.

CHRISTIE, J. M., BLACKWOOD, L., PETERSEN, J., SULLIVAN, S. (2015) Plant
  Flavoprotein Photoreceptors. Plant and Cell Physiology. 56(3):401-413.
}
\keyword{datasets}
