% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phy-Sigma-FR.R
\name{Phy_Sigma_FR}
\alias{Phy_Sigma_FR}
\title{Pfr Sigma as a function of wavelength}
\usage{
Phy_Sigma_FR(w.length, use.cached.mult = FALSE)
}
\arguments{
\item{w.length}{numeric array of wavelength (nm)}

\item{use.cached.mult}{logical ignored}
}
\value{
a numeric array with values for Sigma
}
\description{
Pfr Sigma as a function of wavelength, calculated by interpolation from data
for Type I Phytochrome as compiled by Mancinelli (1994).
}
\examples{
with(sun.spct, Phy_Sigma_FR(w.length))
with(sun.spct, Phy_Sigma_FR(w.length, TRUE))

}
\references{
Mancinelli, A.L. (1994) The physiology of phytochrome action. In
Photomorphogenesis in plants, 2nd edition. R.E. Kendrick and G.H.M.
Kronenberg, eds. Kluwer Academic Publishers, Dordrecht, pp. 211-269. ISBN
978-0-7923-2551-2 (print), 978-94-011-1884-2 (on-line).
\doi{10.1007/978-94-011-1884-2_10}
}
\seealso{
Other phytochrome-related functions and data:
\code{\link{PHYs.mspct}},
\code{\link[=Pfr_Ptot]{Pfr_Ptot()}},
\code{\link[=Pfr_Ptot_R_FR]{Pfr_Ptot_R_FR()}},
\code{\link[=Phy_Sigma]{Phy_Sigma()}},
\code{\link[=Phy_Sigma_R]{Phy_Sigma_R()}},
\code{\link[=Phy_reaction_rates]{Phy_reaction_rates()}}
}
\concept{phytochrome-related functions and data}
