% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytochrome-data.r
\docType{data}
\name{phytochrome.spct}
\alias{phytochrome.spct}
\title{Tabulated data for Phytochrome Sigma}
\format{A generic.spct object with 49 rows and 3 variables}
\description{
A dataset containing the wavelengths at a 1 nm interval.
  Tabulated values for Sigma R and Sigma FR for Type I Phytochrome as
  compiled by Mancinelli (1994).

  The variables are as follows:

  \itemize{ \item wavelength (nm) \item Sigma.R (quantum effectiveness) \item
  Sigma.FR (quantum effectiveness) }
}
\references{
Mancinelli, A.L. (1994) The physiology of phytochrome action. In
Photomorphogenesis in plants, 2nd edition. R.E. Kendrick and G.H.M.
Kronenberg, eds. Kluwer Academic Publishers, Dordrecht, pp. 211-269. ISBN
978-0-7923-2551-2 (print), 978-94-011-1884-2 (on-line).
\href{http://dx.doi.org/10.1007/978-94-011-1884-2_10}{DOI
10.1007/978-94-011-1884-2_10}
}
\keyword{datasets}

