% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQ_redox_state.R
\docType{data}
\name{PS1_PS2_k.spct}
\alias{PS1_PS2_k.spct}
\title{Coefficients for redox state of the plastoquinone pool}
\format{
\code{generic_spct} object with 351 rows and 2 variables.
}
\usage{
PS1_PS2_k.spct
}
\description{
The variables are as follows:
}
\details{
\itemize{
  \item w.length (nm)
  \item k
}
where \code{k} contains fitted coefficients of a model for estimating the
redox state of the plastoquinone pool (Mattila et al., 2020). Based on
measurements in \emph{Arabidopsis} plants. If used in a publication, please,
cite both Mattila et al. (2020) and this R package.
}
\seealso{
\code{\link{PQ_redox_state}()}
}
\keyword{datasets}
