% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs-anatomy.R
\name{rs_from_size}
\alias{rs_from_size}
\alias{gs_from_size}
\alias{gs_w_from_size}
\alias{gs_c_from_size}
\title{Stomatal conductance from sizes and density}
\usage{
rs_from_size(length, width = length, depth, num = 1, D)

gs_from_size(length, width = length, depth, num = 1, D)

gs_w_from_size(length, width = length, depth, num = 1, temperature)

gs_c_from_size(length, width = length, depth, num = 1, temperature)
}
\arguments{
\item{length, width}{numeric "Length" and "width" of the stomatal opening,
i.e., the major and minor diameters of the ellipse describing the stomatal
pore's mean cross section [\eqn{m}].}

\item{depth}{numeric The depth (or diffusion path length) of the stomatal
pore [\eqn{m}].}

\item{num}{numeric Number of stomata per unit area [\eqn{m^{-2}}].}

\item{D}{numeric Diffusion coefficient of the gas [\eqn{m^2\,s^{-1}}].}

\item{temperature}{numeric Leaf temperature. Used to estimate the diffusion
coefficient \code{D}.}
}
\value{
A \code{numeric} vector of resistances expressed in \eqn{s\,m^{-1}}
or \eqn{s\,mol^{-1}}, or of conductances expressed in \eqn{m\,s^{-1}} or in
\eqn{mol\,m^{-2}\,s^{-1}} when the density expressed in estomata per \eqn{m^2}
is passed as argument to \code{n}. With the default of \code{n = 1} the
diffusive conductance per individual pore is returned expressed, e.g., for
molar conductance, in \eqn{mol\,s-1}.
}
\description{
Computations based on mass transfer through pores.
}
\details{
The equation for \eqn{r_s} from Monteith and Unsworth (2008):

\deqn{r_s = \frac{4(l + \pi\,d / 8)}{\pi\,n\,d^2\,D}}

is used to compute diffusive resistance in function \code{rs_from_size()},
which can be used to obtain diffusive resistance from a known size of a
pore with circular or elliptical cross section. The value of \code{D}, the
diffusion should match that of water vapour or \eqn{CO_2}, and its unit
of expression determines the whether the returned value is expressed as the
inverse of a volume or molar flux rate.

Functions \code{gs_from_size()}, \code{gs_w_from_size()} and
\code{gs_c_from_size()} are convenience wrappers.

Arguments are expected in SI units with no scale multiplier. Enter dimensions
in micrometers using scientific notation, e.g., \code{length = 20e-6} for a
pore length of \eqn{20\,\mu m} and \code{n = 10e6} for a density of
\eqn{10\,mm^{-2}}.
}
\note{
This is an approximate computation as plant stomata have a section that
varies with depth. The equation incorporates a single end correction and
assumes no interference among the flows from neighbouring stomata crossing
the bounday layer.
}
\examples{
# a single round stomatal pore
gs_from_size(length = 20e-6, depth = 5e-6, D = D_water(23))

# a single elliptical stomatal pore
gs_from_size(length = 30e-6, width = 10e-6, depth = 5e-6, D = D_water(23))

# 200 circular stomatal pores per mm^2
rs_from_size(length = 5e-6,
             width = 5e-6,
             depth = 10e-6,
             num = 200e6,
             D = D_water(25))

# 50 elliptical stomatal pores per mm^2
rs_from_size(length = 10e-6,
             width = 5-6,
             depth = 20e-6,
             num = 50e6,
             D = D_water(23))

# a single round stomatal pore with dimensions in micrometers
gs_from_size(length = 20, depth = 5, D = D_water(23))

}
\references{
Monteith, J. L. and Unsworth M. H. (2008) Principles of Environmental
  Physics (3ed) Academic Press-Elsevier. ISBN: 978-0-12-505103-3. See
  Section 11.4 Mass transfer through pores.
}
\seealso{
Functions \code{\link{D_water}()} and \code{\link{D_CO2}()} can be
used to compute the diffusion coefficients as a function of temperature. In
addition function \code{\link{molar_vol}()} computes the molar volume of an
ideal gas as a function of temperature and pressure. Functions
\code{\link{gs_mol2vol}()} and \code{\link{gs_vol2mol}()} interconvert
conductances between molar and volume bases of expression. These functions
are used internally in the functions described here.
}
