% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant-qq-ratios.r
\name{UVB_PAR}
\alias{UVB_PAR}
\title{Calculate UVB:PAR photon ratio from spectral irradiance.}
\usage{
UVB_PAR(spct, std = "ISO", use.cached.mult = FALSE, use.hinges = TRUE)
}
\arguments{
\item{spct}{an object of class "source.spct".}

\item{std}{select which definition of UVB should be used,
defaults to "ISO".}

\item{use.cached.mult}{logical indicating whether multiplier values should be
cached between calls.}

\item{use.hinges}{logical indicating whether to use hinges to reduce
interpolation errors.}
}
\value{
a single numeric dimensionless value giving the UVB:UV photon ratio,
  with name attribute set to the name of the wavebands, with "(q:q)"
  appended.
}
\description{
This function returns the UVB:PAR photon ratio of a light source
spectrum.
}
\examples{
UVB_PAR(sun.spct)

}
\seealso{
\code{\link[photobiologyWavebands]{UVB}},
  \code{\link[photobiologyWavebands]{PAR}} and
  \code{\link[photobiology]{q_ratio}}.
}
