% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztls-mspct.r
\docType{data}
\name{ZTLs.mspct}
\alias{ZTLs.mspct}
\title{ZTL absorbance spectra.}
\format{A \code{filter_mspct} with five member \code{filter_spct} objects each
  with 300 rows and 2 numeric variables, \code{w.length} and \code{A}}
\description{
A dataset containing the wavelengths at an arbitrary nm interval. Tabulated
values for the in vitro absorbance spectrum of ZTL LOV2 domain from
Arabidopsis measured in vitro. Data were digitized from figure
2B in Zoltowski and Imaizumi (2014).
}
\details{
The variables of the member spectra are as follows:

\itemize{ \item w.length (nm) \item A (spectral absorbance) }
}
\note{
If you use these data in a publication, please cite also the original
  source as given under references.
}
\references{
Zoltowski, B. D., Imaizumi, T. (2014). Structure and Function of the
  ZTL/FKF1/LKP2 Group Proteins in Arabidopsis. Enzymes, 35, 213-39.
}
\keyword{datasets}
