% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCree-photosynthesis-mspct.r
\docType{data}
\name{McCree_photosynthesis.mspct}
\alias{McCree_photosynthesis.mspct}
\title{McCree's action spectra for whole-leaf photosynthesis.}
\format{A \code{response_mspct} object with two member \code{response_spct}
  objects each with 200 rows and 2 numeric variables, \code{w.length} and
  \code{s.e.response}.}
\description{
The 'classical' action spectra of K. J. McCree (1972) for
  Amaranthus edulis Speg. var. UCD 1966 and Avena sativa L. var. Coronado are
  included in this data set. Response is net CO2 uptake measured of leaf
  sections. Light source was a xenon-arc lamp fitted with a monochromator.
  Irradiance was in the range 10 to 15 W m$^{-2}$.
}
\note{
Digitised from bitmap of from the original publicatiom.
}
\references{
McCree, K. J. (1972) Significance of Enhancement for Calculations
  Based on the Action Spectrum for Photosynthesis. Plant Physiology, 49,
  704-706. Fig. 1, AMARANTH
}
\keyword{datasets}
