% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQ_redox_state.R
\name{PQ_redox_state}
\alias{PQ_redox_state}
\title{Redox state of the plastoquinone pool}
\usage{
PQ_redox_state(light.spct, coefs.spct = photobiologyPlants::PS1_PS2_k.spct)
}
\arguments{
\item{light.spct}{A \code{source_spct} object.}

\item{coefs.spct}{A \code{generic_spct} object containg the coefficients of
the fitted model. Defaults to those in Mattila et al. (2020).}
}
\value{
A numeric value.
}
\description{
Estimate the redox state of the plastoquinone pool from spectral irradiance.
}
\details{
The computations follow the procedure in Mattila et al. (2020). The
  spectrum in \code{light.spct} is re-expressed by interpolation to the
  wavelengths from \code{coefs.spct}. If used in a publication, please, cite
  both Mattila et al. (2020) and this R package.
}
\examples{

PQ_redox_state(sun.spct)
PQ_redox_state(white_led.source_spct)

}
\references{
Mattila H, Khorobrykh S, Hakala-Yatkin M, Havurinne V, Kuusisto I, Antal T,
Tyystjärvi T, Tyystjärvi E. 2020. Action spectrum of the redox state of the
plastoquinone pool defines its function in plant acclimation. The Plant
Journal 104, 1088–1104. \doi{10.1111/tpj.14983}.
}
