% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs-anatomy.R
\name{rs_from_size}
\alias{rs_from_size}
\alias{gs_from_size}
\alias{gs_w_from_size}
\alias{gs_c_from_size}
\title{Stomatal conductance from sizes and density}
\usage{
rs_from_size(length, width = length, depth, num = 1, D)

gs_from_size(length, width = length, depth, num = 1, D)

gs_w_from_size(length, width = length, depth, num = 1, temperature)

gs_c_from_size(length, width = length, depth, num = 1, temperature)
}
\arguments{
\item{length, width}{numeric "Length" and "width" of the stomatal opening,
i.e., the major and minor diameters of the ellipse describing the stomatal
pore's mean cross section [\eqn{\mu m}].}

\item{depth}{numeric The depth (or diffusion path length) of the stomatal
pore [\eqn{\mu m}].}

\item{num}{numeric Number of stomata per unit area [\eqn{mm^{-2}}].}

\item{D}{numeric Diffusion coefficient of the gas[\eqn{m^2\,s^{-1}}].}

\item{temperature}{numeric Leaf temperature. Used to estimate the diffusion
coefficient \code{D}.}
}
\value{
A resistance expressed in \eqn{s m^{-1}} or a conductance
  expressed in \eqn{m s^{-1}}.
}
\description{
Computations based on mass transfer through pores.
}
\examples{
# a single round stomatal pore
gs_from_size(length = 20, depth = 5, D = D_water(23))
# a single elliptical stomatal pore
gs_from_size(length = 30, width = 10, depth = 5, D = D_water(23))
#
# 200 circular stomatal pores per mm^2
rs_from_size(length = 5,
             width = 5,
             depth = 10,
             num = 200e6,
             D = D_water(25))
# 50 elliptical stomatal pores per mm^2
rs_from_size(length = 10,
             width = 5,
             depth = 20,
             num = 50e6,
             D = D_water(23))

}
\references{
Monteith, J. L. and Unsworth M. H. (2008) Principles of Environmental
  Physics (3ed) Academic Press-Elsevier. ISBN: 978-0-12-505103-3. See
  Section 11.4 Mass transfer through pores.
}
